telegram-bot/config.py
# config.py
# Токени, URI та інші конфігурації з .env.
from dotenv import load_dotenv
import os

load_dotenv()

BOT_TOKEN = os.getenv("BOT_TOKEN")
MONGO_URI = os.getenv("MONGO_URI")
